+++ plugins/wasm-test/src/lib.rs
@@ -26,6 +26,65 @@ struct WasmTestConfig {
 }
 
 #[plugin_fn]
+pub fn testing_macros(_: ()) -> FnResult<()> {
+    // Errors
+    let _ = plugin_err!(code = 2, "Error");
+    let _ = plugin_err!(code = 3, "Error {}", "arg");
+    let _ = plugin_err!("Error");
+    let _ = plugin_err!("Error {}", "arg");
+
+    // Commands
+    let args = ["a", "b", "c"];
+
+    exec_command!("git");
+    exec_command!("git", args);
+    exec_command!("git", ["a", "b", "c"]);
+    exec_command!(input, ExecCommandInput::default());
+    exec_command!(pipe, "git");
+    exec_command!(pipe, "git", args);
+    exec_command!(pipe, "git", ["a", "b", "c"]);
+    exec_command!(inherit, "git");
+    exec_command!(inherit, "git", args);
+    exec_command!(inherit, "git", ["a", "b", "c"]);
+    let _ = exec_command!(raw, ExecCommandInput::default());
+    let _ = exec_command!(raw, "git");
+    let _ = exec_command!(raw, "git", args);
+    let _ = exec_command!(raw, "git", ["a", "b", "c"]);
+
+    // Env vars
+    let name = "VAR";
+
+    let _ = host_env!("VAR");
+    let _ = host_env!(name);
+    host_env!("VAR", "value");
+    host_env!("VAR", name);
+    host_env!(name, name);
+    host_env!(name, "value");
+
+    // Logging
+    host_log!("Message");
+    host_log!("Message {} {} {}", 1, 2, 3);
+    host_log!(input, HostLogInput::default());
+    host_log!(stdout, "Message");
+    host_log!(stdout, "Message {} {} {}", 1, 2, 3);
+    host_log!(stderr, "Message");
+    host_log!(stderr, "Message {} {} {}", 1, 2, 3);
+
+    // Paths
+    let path = "/proto/path";
+    let pathbuf = PathBuf::from("/proto/buf");
+
+    let _ = real_path!("/proto/dir");
+    let _ = real_path!(path);
+    let _ = real_path!(buf, pathbuf);
+    let _ = virtual_path!("/proto/dir");
+    let _ = virtual_path!(path);
+    let _ = virtual_path!(buf, pathbuf);
+
+    Ok(())
+}
+
+#[plugin_fn]
 pub fn register_tool(_: ()) -> FnResult<Json<ToolMetadataOutput>> {
     host_log!(stdout, "Registering tool");